File opening checks in upload.cgi

+++ www/cgi-bin/upload.cpp
@@ -6,44 +6,57 @@
 
 int ft_stoi(std::string);
 std::string successPage();
+std::string errorPage();
 std::string getFilename(std::string &);
 std::string getBoundary(std::string &);
 int findStartOfFile(std::string &, std::string &);
 
 int main(int argc, char **argv, char **env)
 {
-	std::ifstream infile("tmp", std::ifstream::in | std::ifstream::binary);
+	std::ifstream infile;
 	std::ofstream outfile;
 	std::string file, filename, boundary, message;
 	std::string socket(argv[0]);
 	char c;
 	int pos, sock;
 
-	while (true)
+	infile.open("tmp", std::ifstream::in | std::ifstream::binary);
+
+	if (infile.fail())
+		message = errorPage();
+	else
 	{
-		if (infile.eof())
-			break;
-		infile.read(&c, 1);
-		file.append(1, c);
+		while (true)
+		{
+			if (infile.eof())
+				break;
+			infile.read(&c, 1);
+			file.append(1, c);
+		}
+		infile.close();
+
+		sock = ft_stoi(&socket[socket.find_first_of("=") + 1]);
+
+		boundary = getBoundary(file);
+		boundary.insert(0, "--");
+		filename = getFilename(file);
+		filename.insert(0, "www/uploads/");
+		pos = findStartOfFile(file, boundary);
+		file.erase(0, pos);
+		file.erase(file.find(boundary) - 4, file.npos);
 	}
-	infile.close();
-
-	sock = ft_stoi(&socket[socket.find_first_of("=") + 1]);
-
-	boundary = getBoundary(file);
-	boundary.insert(0, "--");
-	filename = getFilename(file);
-	filename.insert(0, "www/uploads/");
-	pos = findStartOfFile(file, boundary);
-	file.erase(0, pos);
-	file.erase(file.find(boundary) - 4, file.npos);
 
 	outfile.open(filename, std::ofstream::out | std::ofstream::trunc | std::ofstream::binary);
-	outfile << file;
-	outfile.close();
-	unlink("tmp");
 
-	message = successPage();
+	if (outfile.fail())
+		message = errorPage();
+	else
+	{
+		outfile << file;
+		outfile.close();
+		unlink("tmp");
+		message = successPage();
+	}
 
 	send(sock, message.c_str(), message.size(), 0);
 
@@ -91,6 +104,31 @@ std::string successPage()
 	             "<body>\n"
 	             "<center><h1>"
 				 "You file has been successfully uploaded !"
+	      "</h1></center>\n"
+	             "</body>\n"
+	             "</html>";
+	html = ss.str();
+	message.insert(message.find("\r\nContent-Type"), ft_to_string(html.size()));
+	message.append(html);
+	return message;
+}
+
+std::string successPage()
+{
+	std::string html, message = "HTTP/1.1 500 Internal Server Error\r\nConnection: keep-alive\r\nContent-Length: \r\nContent-Type: text/html\r\n\r\n";
+	std::stringstream ss;
+	ss << "<!DOCTYPE html>"
+	      "<html>\n"
+	      "<head>"
+		  	"<meta charset=\"UTF-8\">\n"
+   		  	"<meta name=\"viewport\" content=\"width=device-width, initial-scale=1.0\">\n"
+		  	"<link rel=\"shortcut icon\" type=\"image/x-icon\" href=\"/assets/favicon.ico\">\n"
+    	  	"<link rel=\"stylesheet\" type=\"text/css\" href=\"/assets/stylesheet.css\">\n"
+		  	"<title>Upload failed</title>\n"
+	      "</head>\n"
+	             "<body>\n"
+	             "<center><h1>"
+				 "500 Internal Server Error"
 	      "</h1></center>\n"
 	             "</body>\n"
 	             "</html>";
